test(all): extract common jest config
