+++ apps/browser-extension-wallet/test/jest.config.js
@@ -1,27 +1,8 @@
 const { pathsToModuleNameMapper } = require('ts-jest');
 const { compilerOptions } = require('../src/tsconfig');
-const { jestEsmExceptions } = require('../../../test/jestEsmExceptions');
+const { createJestConfig } = require('../../../test/createJestConfig');
 
-//Modules, that are loaded in esm versions and require explicit transformation
-const esmExceptions = jestEsmExceptions([
-  'style-inject',
-  'tslib',
-  '@cardano-ogmios',
-  'nanoid',
-  'rxjs',
-  'esm-browser',
-  'uuid',
-  '@react-rxjs',
-  'dexie',
-  '@rxstate/core',
-  'intersection-observer-polyfill',
-  'p-retry'
-]);
-
-const rootDir = process.cwd();
-
-module.exports = {
-  rootDir,
+module.exports = createJestConfig({
   moduleNameMapper: {
     '.*\\.(scss|sass|css|less)$': '<rootDir>/test/__mocks__/styleMock.js',
     '.*\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2)$': '<rootDir>/test/__mocks__/fileMock.js',
@@ -32,15 +13,7 @@ module.exports = {
     uuid: require.resolve('uuid'),
     ...pathsToModuleNameMapper(compilerOptions.paths, { prefix: '<rootDir>/src' })
   },
-  preset: 'ts-jest',
   roots: ['<rootDir>/src'],
-  transform: {
-    '^.+\\.test.ts?$': 'ts-jest',
-    '^.+\\.test.tsx?$': 'ts-jest',
-    '^.+\\.(ts|tsx)$': 'ts-jest',
-    ...esmExceptions.transform
-  },
-  transformIgnorePatterns: esmExceptions.transformIgnorePatterns,
   testTimeout: 60000,
   testEnvironment: 'jsdom',
   setupFiles: [
@@ -78,4 +51,4 @@ module.exports = {
       tsconfig: './src/tsconfig.json'
     }
   }
-};
+});
