+++ packages/cardano/test/jest.config.js
@@ -1,21 +1,10 @@
 const { pathsToModuleNameMapper } = require('ts-jest');
 const { compilerOptions } = require('../src/tsconfig');
-const { jestEsmExceptions } = require('../../../test/jestEsmExceptions');
+const { createJestConfig } = require('../../../test/createJestConfig');
 
 const rootDir = process.cwd();
-//Modules, that are loaded in esm versions and require explicit transformation
-const esmExceptions = jestEsmExceptions([
-  'style-inject',
-  'tslib',
-  '@cardano-ogmios',
-  'nanoid',
-  'rxjs',
-  'esm-browser',
-  'uuid'
-]);
 
-module.exports = {
-  rootDir,
+module.exports = createJestConfig({
   moduleNameMapper: {
     '.*\\.(scss|sass|css|less)$': '<rootDir>/test/__mocks__/styleMock.js',
     '.*\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2)$': '<rootDir>/test/__mocks__/fileMock.js',
@@ -27,14 +16,7 @@ module.exports = {
     uuid: require.resolve('uuid'),
     ...pathsToModuleNameMapper(compilerOptions.paths, { prefix: '<rootDir>/src' })
   },
-  preset: 'ts-jest',
   roots: ['<rootDir>/src'],
-  transform: {
-    '^.+\\.test.ts?$': 'ts-jest',
-    '^.+\\.test.tsx?$': 'ts-jest',
-    ...esmExceptions.transform
-  },
-  transformIgnorePatterns: esmExceptions.transformIgnorePatterns,
   testTimeout: 60000,
   testEnvironment: 'jsdom',
   collectCoverageFrom: [
@@ -50,4 +32,4 @@ module.exports = {
       tsconfig: './src/tsconfig.json'
     }
   }
-};
+});
